region-app/requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
requests==2.24.0
cachetools==4.1.1
gunicorn==20.0.4
